ml_backend/requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.0
numpy>=2.0.0
